# backend/core/policy_engine/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Grandway Policy-Registered API",
    "version": "1.0.0",
    "description": "Generated from the Core Policy Engine registry (source of truth for endpoints, permission keys, HTTP methods, and risk levels). Request/response body schemas are documented per app in `docs/DATA_CONTRACT.md`."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/applicants/{applicant_id}/status/": {
      "post": {
        "operationId": "applicants.applicant.change_status",
        "summary": "Change Applicant Status",
        "description": "Set the applicant's standing: active, dormant, or archived. Always manual.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicants.applicant.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "applicant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/document-templates/signatories/{signatory_id}/status/": {
      "post": {
        "operationId": "document_templates.signatory.change_status",
        "summary": "Change Signatory Status",
        "description": "Move a signatory between draft, active, and inactive. Nothing is deleted.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "document_templates"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "document_templates.signatory.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "signatory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/document-templates/templates/{template_id}/status/": {
      "post": {
        "operationId": "document_templates.template.change_status",
        "summary": "Change Document Template Status",
        "description": "Move a template between draft, active, and inactive. Existing documents are unaffected.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "document_templates"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "document_templates.template.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/documents/{document_id}/status/": {
      "post": {
        "operationId": "documents.document.change_status",
        "summary": "Change Document Status",
        "description": "Move a document between draft and ready. Cannot set archived.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "documents.document.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/offers/conditions/{condition_id}/status/": {
      "post": {
        "operationId": "offers.condition.change_status",
        "summary": "Change Offer Condition Status",
        "description": "Mark a condition satisfied, waived, not applicable, or pending again.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "offers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "offers.condition.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "condition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "description": "Resource payload — see the owning app's DATA_CONTRACT.md."
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "APP_RESOURCE_REASON code."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "Meta": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "next": {
            "type": [
              "string",
              "null"
            ]
          },
          "previous": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Standard error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}